+++ modules/pag/__init__.py
@@ -14,7 +14,7 @@ def apply(p: processing.StableDiffusionProcessing): # pylint: disable=arguments-
     cls = shared.sd_model.__class__ if shared.sd_loaded else None
     if cls == StableDiffusionPAGPipeline or cls == StableDiffusionXLPAGPipeline:
         cls = unapply()
-    if p.pag_scale == 0:
+    if p.cfg_true == 0:
         return
     if cls is not None and 'PAG' in cls.__name__:
         pass
@@ -31,20 +31,20 @@ def apply(p: processing.StableDiffusionProcessing): # pylint: disable=arguments-
         orig_pipeline = shared.sd_model
         shared.sd_model = sd_models.switch_pipe(StableDiffusionXLPAGPipeline, shared.sd_model)
     elif detect.is_f1(cls):
-        p.task_args['true_cfg_scale'] = p.pag_scale
+        p.task_args['true_cfg_scale'] = p.cfg_true
     else:
         # log.warning(f'PAG: pipeline={cls.__name__} required={StableDiffusionPipeline.__name__}')
         return None
 
-    p.task_args['pag_scale'] = p.pag_scale
-    p.task_args['pag_adaptive_scaling'] = p.pag_adaptive
-    p.task_args['pag_adaptive_scale'] = p.pag_adaptive
+    p.task_args['cfg_true'] = p.cfg_true
+    p.task_args['cfg_adaptive_scaling'] = p.cfg_adaptive
+    p.task_args['cfg_adaptive_scale'] = p.cfg_adaptive
     pag_applied_layers = shared.opts.pag_apply_layers
     pag_applied_layers_index = pag_applied_layers.split() if len(pag_applied_layers) > 0 else []
     pag_applied_layers_index = [p.strip() for p in pag_applied_layers_index]
     p.task_args['pag_applied_layers_index'] = pag_applied_layers_index if len(pag_applied_layers_index) > 0 else ['m0'] # Available layers: d[0-5], m[0], u[0-8]
-    p.extra_generation_params["CFG true"] = p.pag_scale
-    p.extra_generation_params["CFG adaptive"] = p.pag_adaptive
+    p.extra_generation_params["CFG true"] = p.cfg_true
+    p.extra_generation_params["CFG adaptive"] = p.cfg_adaptive
     # log.debug(f'{c}: args={p.task_args}')