+++ modules/pag/pipe_sd.py
@@ -1034,20 +1034,20 @@ class StableDiffusionPAGPipeline(
         return self._interrupt
 
     @property
-    def pag_scale(self):
-        return self._pag_scale
+    def cfg_true(self):
+        return self._cfg_true
 
     @property
     def do_perturbed_attention_guidance(self):
-        return self._pag_scale > 0
+        return self._cfg_true > 0
 
     @property
-    def pag_adaptive_scaling(self):
-        return self._pag_adaptive_scaling
+    def cfg_adaptive_scaling(self):
+        return self._cfg_adaptive_scaling
 
     @property
-    def do_pag_adaptive_scaling(self):
-        return self._pag_adaptive_scaling > 0
+    def do_cfg_adaptive_scaling(self):
+        return self._cfg_adaptive_scaling > 0
 
     @property
     def pag_applied_layers_index(self):
@@ -1063,8 +1063,8 @@ class StableDiffusionPAGPipeline(
         num_inference_steps: int = 50,
         timesteps: List[int] | None = None,
         guidance_scale: float = 7.5,
-        pag_scale: float = 0.0,
-        pag_adaptive_scaling: float = 0.0,
+        cfg_true: float = 0.0,
+        cfg_adaptive_scaling: float = 0.0,
         pag_applied_layers_index: List[str] = ["d4"],  # ['d4', 'd5', 'm0']
         negative_prompt: Optional[Union[str, List[str]]] = None,
         num_images_per_prompt: Optional[int] = 1,
@@ -1202,8 +1202,8 @@ class StableDiffusionPAGPipeline(
         self._cross_attention_kwargs = cross_attention_kwargs
         self._interrupt = False
 
-        self._pag_scale = pag_scale
-        self._pag_adaptive_scaling = pag_adaptive_scaling
+        self._cfg_true = cfg_true
+        self._cfg_adaptive_scaling = cfg_adaptive_scaling
         self._pag_applied_layers_index = pag_applied_layers_index
 
         # 2. Define call parameters
@@ -1372,9 +1372,9 @@ class StableDiffusionPAGPipeline(
                 elif not self.do_classifier_free_guidance and self.do_perturbed_attention_guidance:
                     noise_pred_original, noise_pred_perturb = noise_pred.chunk(2)
 
-                    signal_scale = self.pag_scale
-                    if self.do_pag_adaptive_scaling:
-                        signal_scale = self.pag_scale - self.pag_adaptive_scaling * (1000 - t)
+                    signal_scale = self.cfg_true
+                    if self.do_cfg_adaptive_scaling:
+                        signal_scale = self.cfg_true - self.cfg_adaptive_scaling * (1000 - t)
                         if signal_scale < 0:
                             signal_scale = 0
 
@@ -1384,9 +1384,9 @@ class StableDiffusionPAGPipeline(
                 elif self.do_classifier_free_guidance and self.do_perturbed_attention_guidance:
                     noise_pred_uncond, noise_pred_text, noise_pred_text_perturb = noise_pred.chunk(3)
 
-                    signal_scale = self.pag_scale
-                    if self.do_pag_adaptive_scaling:
-                        signal_scale = self.pag_scale - self.pag_adaptive_scaling * (1000 - t)
+                    signal_scale = self.cfg_true
+                    if self.do_cfg_adaptive_scaling:
+                        signal_scale = self.cfg_true - self.cfg_adaptive_scaling * (1000 - t)
                         if signal_scale < 0:
                             signal_scale = 0