+++ modules/video_models/video_run.py
@@ -46,7 +46,7 @@ def generate(*args, **kwargs):
         denoising_strength=float(init_strength),
         init_image=init_image,
         cfg_scale=float(guidance_scale),
-        pag_scale=float(guidance_true),
+        cfg_true=float(guidance_true),
         vae_type=vae_type,
         vae_tile_frames=int(vae_tile_frames),
         override_settings=override_settings,